[CL-3197] Use the same Modal
