+++ front/app/components/admin/SeatChangeSuccess/messages.ts
