+++ front/app/containers/Admin/users/ChangeSeatModal.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 // Components
 import { Box, Button, Text } from '@citizenlab/cl2-component-library';
 import Modal from 'components/UI/Modal';
-import SeatChangeSuccessModal from 'components/admin/SeatChangeSuccessModal';
+import SeatChangeSuccess from 'components/admin/SeatChangeSuccess';
 
 // Translation
 import { FormattedMessage, MessageDescriptor, useIntl } from 'utils/cl-intl';
@@ -86,67 +86,65 @@ const ChangeSeatModal = ({
     hasExceededSetAdmins,
     hasSeatBasedBillingEnabled
   );
+  const header = !showSuccessModal ? (
+    <Box px="2px">
+      <Text color="primary" my="8px" fontSize="l" fontWeight="bold">
+        {formatMessage(modalTitle)}
+      </Text>
+    </Box>
+  ) : undefined;
 
   return (
     <>
-      <Modal
-        opened={showModal}
-        close={closeModal}
-        header={
-          <Box px="2px">
-            <Text color="primary" my="8px" fontSize="l" fontWeight="bold">
-              {formatMessage(modalTitle)}
-            </Text>
-          </Box>
-        }
-      >
-        <Box display="flex" flexDirection="column" width="100%" p="32px">
-          <Box>
-            <Text color="textPrimary" fontSize="m" my="0px">
-              <FormattedMessage
-                {...confirmChangeQuestion}
-                values={{
-                  name: (
-                    <Text as="span" fontWeight="bold" fontSize="m">
-                      {`${userToChangeSeat.attributes.first_name} ${userToChangeSeat.attributes.last_name}`}
-                    </Text>
-                  ),
+      <Modal opened={showModal} close={closeModal} header={header}>
+        {showSuccessModal && (
+          <Box display="flex" flexDirection="column" width="100%" p="32px">
+            <Box>
+              <Text color="textPrimary" fontSize="m" my="0px">
+                <FormattedMessage
+                  {...confirmChangeQuestion}
+                  values={{
+                    name: (
+                      <Text as="span" fontWeight="bold" fontSize="m">
+                        {`${userToChangeSeat.attributes.first_name} ${userToChangeSeat.attributes.last_name}`}
+                      </Text>
+                    ),
+                  }}
+                />
+              </Text>
+              <Box py="32px">
+                <SeatInfo seatType="admin" />
+              </Box>
+            </Box>
+            <Box
+              display="flex"
+              flexDirection="row"
+              width="100%"
+              alignItems="center"
+            >
+              <Button
+                width="auto"
+                onClick={() => {
+                  toggleAdmin();
+                  // closeModal();
+                  if (!isUserAdmin) {
+                    setShowSuccessModal(true);
+                  }
                 }}
-              />
-            </Text>
-            <Box py="32px">
-              <SeatInfo seatType="admin" />
+              >
+                {formatMessage(buttonText)}
+              </Button>
             </Box>
           </Box>
-          <Box
-            display="flex"
-            flexDirection="row"
-            width="100%"
-            alignItems="center"
-          >
-            <Button
-              width="auto"
-              onClick={() => {
-                toggleAdmin();
-                closeModal();
-                if (!isUserAdmin) {
-                  setShowSuccessModal(true);
-                }
-              }}
-            >
-              {formatMessage(buttonText)}
-            </Button>
-          </Box>
-        </Box>
+        )}
+        {!isUserAdmin && showSuccessModal && (
+          <SeatChangeSuccess
+            closeModal={closeModal}
+            hasPurchasedMoreSeats={hasExceededSetAdmins}
+            seatType="admin"
+          />
+        )}
       </Modal>
-      {!isUserAdmin && (
-        <SeatChangeSuccessModal
-          showModal={showSuccessModal}
-          closeModal={() => setShowSuccessModal(false)}
-          hasPurchasedMoreSeats={hasExceededSetAdmins}
-          seatType="admin"
-        />
-      )}
     </>
   );
 };
